requirements.txt
pyserial>=3.5
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scipy>=1.7.0
configparser>=5.0.0